Do some reorganising and cleaning
